Rename a common cpp header file
