+++ cpp/common/common.h
