+++ cpp/filtering/filtering.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "utils.h"
+#include "common.h"
 
 DLL_API void intensity_capping(Byte* array, unsigned int array_size, double n_std);
