+++ cpp/pooling/pooling.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "utils.h"
+#include "common.h"
 
 DLL_API float mag_pool(float* array, int size, float k);
